tianshu_serving/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
black==19.10b0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
dataclasses==0.8
fastapi==0.61.2
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h11==0.11.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MarkupSafe==1.1.1
mock==4.0.2
numpy==1.19.4
oauthlib==3.1.0
oneflow-cu102==0.2b2
onnx==1.8.0
opt-einsum==3.3.0
pandas==1.1.5
pathspec==0.8.1
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.2
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
starlette==0.13.6
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
toml==0.10.2
torch @ file:///root/torch-1.6.0-cp36-cp36m-linux_x86_64.whl
torchvision @ file:///root/torchvision-0.7.0-cp36-cp36m-manylinux1_x86_64.whl
tqdm==4.53.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.12.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
